services/ai-agent/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
nats-py==2.7.2
langchain==0.2.1
langchain-anthropic==0.1.13
langchain-community==0.2.1
ollama==0.2.1
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.32.2
